+++ tgui/packages/tgui/interfaces/Uplink/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_SEARCH_RESULTS = 25;
